BigFloat compatibility for univariate optimization, fixes #271 (#336)
* BigFloat compatibility for univariate optimization, fixes #271

The univariate optimization code unintentionally assumed that the choice
variable was a Float64.  Specifically, the interfaces accommodated an
AbstractFloat, but then would crash when reporting the results because of
missing typecasts.

This patch introduces approriate typecasts and test cases.

* fix typecasts -- needed for abs_tol, rel_tol (not lower and upper)

* refactor univariate tests

* drop all Float64 fields from UnivariateOptimizationResults

+++ src/brent.jl
@@ -158,7 +158,7 @@ function optimize{F <: Function, T <: AbstractFloat}(
                                          initial_lower,
                                          initial_upper,
                                          new_minimizer,
-                                         Float64(new_minimum),
+                                         new_minimum,
                                          iteration,
                                          iteration == iterations,
                                          converged,
